LPS-81335 remove structure with form instance

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/exportimport/staged/model/repository/DDMFormInstanceStagedModelRepository.java
@@ -195,7 +195,11 @@ public class DDMFormInstanceStagedModelRepository
 		throws PortalException {
 
 		for (Long ddmStructureId : ddmStructureIds) {
-			_ddmStructureLocalService.deleteStructure(ddmStructureId);
+			if (_ddmStructureLocalService.fetchDDMStructure(ddmStructureId) !=
+					null) {
+
+				_ddmStructureLocalService.deleteStructure(ddmStructureId);
+			}
 		}
 	}
